Prodution

+++ resources/views/documents/advices.blade.php
@@ -124,7 +124,9 @@
                                             <td > {{ $processing->statusProcessingDocument->name ?? '' }}</td>
                                             <td > {{ $processing->destination->name ?? '' }}</td>
                                             <td style="text-align: justify;"> {!! $processing->observation !!}</td>
-                                            <td> <button type="button" class="btn btn-danger btn-xs" onclick="delete_processing('{{$processing->id}}')"> <i class="fa fa-trash"></i> </button> </td>
+                                            @if(Auth::user()->id == $processing->owner->id)
+                                                <td> <button type="button" class="btn btn-danger btn-xs" onclick="delete_processing('{{$processing->id}}')"> <i class="fa fa-trash"></i> </button> </td>
+                                            @endif
                                         </tr>
                                     @empty
                                         <tr>
@@ -185,6 +187,7 @@
                     }).success(function (data) {
                         if (data) {
                             toastr.success("Pedido salvo com sucesso!!");
+                            window.location.reload()
                         } else {
                             toastr.error("Erro ao salvar pedido!!");
                         }
@@ -268,9 +271,12 @@
                             str += "<td>";
                             str += valor.observation || '';
                             str += "</td>";
+                            @if(Auth::user()->id == $processing->owner->id)
                             str += "<td>";
                             str += '<button type="button" class="btn btn-danger btn-xs" onclick="delete_processing(' + valor.id + ')"> <i class="fa fa-trash"></i> </button>';
                             str += "</td>";
+                            @endif
+
                             str += "</tr>";
                             table.append(str);
                         });
@@ -281,6 +287,8 @@
                         $('#new_status_processing_document_id').val(0);
                         $('#new_processing_document_date').val('');
                         CKEDITOR.instances.new_document_observation.setData('');
+
+                        window.location.href = '{{route('documents.index')}}';
                     });
                 }
             }
